+++ common/src/main/java/com/nemonotfound/nemos/copper/item/ModCreativeModeTabs.java
@@ -30,6 +30,10 @@ public class ModCreativeModeTabs {
                 output.accept(ModItems.COPPER_HOE.get());
                 output.accept(ModItems.COPPER_SHEARS.get());
                 output.accept(ModItems.COPPER_SWORD.get());
+                output.accept(ModItems.COPPER_HELMET.get());
+                output.accept(ModItems.COPPER_CHESTPLATE.get());
+                output.accept(ModItems.COPPER_LEGGINGS.get());
+                output.accept(ModItems.COPPER_BOOTS.get());
             })
             .build());