+++ common/src/main/java/com/nemonotfound/nemos/copper/tag/ModItemTags.java
@@ -9,7 +9,8 @@ import static com.nemonotfound.nemos.copper.Constants.MOD_ID;
 
 public class ModItemTags {
 
-    public static final TagKey<Item> COPPER_TOOL_MATERIALS = TagKey.create(Registries.ITEM, ResourceLocation.fromNamespaceAndPath(MOD_ID, "copper_tool_materials.json"));
+    public static final TagKey<Item> COPPER_TOOL_MATERIALS = TagKey.create(Registries.ITEM, ResourceLocation.fromNamespaceAndPath(MOD_ID, "copper_tool_materials"));
+    public static final TagKey<Item> REPAIRS_COPPER_ARMOR = TagKey.create(Registries.ITEM, ResourceLocation.fromNamespaceAndPath(MOD_ID, "repairs_copper_armor"));
 
     public static void init() {}
 }