+++ fabric/src/main/java/com/nemonotfound/nemos/copper/datagen/ModBlockTagProvider.java
@@ -1,0 +1,25 @@
+package com.nemonotfound.nemos.copper.datagen;
+
+import com.nemonotfound.nemos.copper.block.ModBlocks;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricTagProvider;
+import net.minecraft.core.HolderLookup;
+import net.minecraft.tags.BlockTags;
+
+import java.util.concurrent.CompletableFuture;
+
+public class ModBlockTagProvider extends FabricTagProvider.BlockTagProvider {
+
+    public ModBlockTagProvider(FabricDataOutput output, CompletableFuture<HolderLookup.Provider> completableFuture) {
+        super(output, completableFuture);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider wrapperLookup) {
+        getOrCreateTagBuilder(BlockTags.RAILS)
+                .add(ModBlocks.COPPER_RAIL.get())
+                .add(ModBlocks.COPPER_ACTIVATOR_RAIL.get())
+                .add(ModBlocks.COPPER_DETECTOR_RAIL.get())
+                .add(ModBlocks.COPPER_POWERED_RAIL.get());
+    }
+}